new files

+++ Chapter 6/6_1.c
@@ -1,17 +1,17 @@
 #include <stdio.h>
 int main() {
-  float input, largest = 0.0f;
+  int input, largest = 0;
   printf("This program will return the largest number in a series\n");
   printf("Enter a series of numbers, input zero or a negative number to stop\n");
   for (;;){
     printf("Enter a number: ");
-    scanf("%f", &input);
+    scanf("%d", &input);
     if (input < 0 || input == 0)
       break;
     else if (input > largest)
       largest = input;
   }
-  printf("The largest number is %.2f", largest);
+  printf("The largest number is %d", largest);
 
 
   return 0;
